Refactor security architecture
Move authorization handling to a separate layer between controllers and
(internal) services. This layer uses new decorator services which
implement the interface of the entity service and a new `SecuredService`
interface for tagging.

Known issue:
The DefaultEntityServiceImplTest#testCaching fails if it is not run
isolated and has been disabled for now. The issue seems to be caused by
`JsonViewControllerAdviceTest`.